+++ CHANGELOG.md
@@ -14,5 +14,9 @@ Fix bug in date parsing
 Fix connection to MongoDB: use direct connection and secondaryPreferred to avoid issues in certain clusters
 Reorganize packages
 Create and document Dockerfile
-## 0.1.5 wip
+## 0.1.5
 Fix a bug in date parsing
+
+Allow "value" key in the fields definition to override the field with a constant value.
+The field will not be queried, but will be filled with the constant value. 
+E.g., can be set to 'null' if a field is missing the the DB but still needed in the datalake.